+++ .gitignore
@@ -1,1 +1,1 @@
-MarketWatch/secrets.py
+MarketWatch/mw_secrets.py
