+++ MarketWatch/market_watch.py
@@ -1,8 +1,8 @@
 import requests
 import json
-from item_list import *
-from testdata import *
-from secrets import *
+from mw_item_list import *
+from mw_test_data import *
+from mw_secrets import *
 
 def itemInfo(item):
     api_request = API_BASE.format(ITEM_ID=item, SELECTIONS=SELECTIONS, API_KEY=API_KEY)
